feat: Add MRU caching system
This commit adds a new class `MRUCache` that implements a MRU (Most Recently Used) caching system. It extends the `BaseCaching` class and includes the `put` method to add items to the cache and the `get` method to retrieve items by key. The `MRUCache` class maintains a maximum number of items in the cache and discards the most recently used item when the cache is full.