feat: Add LFU caching system
This commit adds a new class `LFUCache` that implements an LFU (Least Frequently Used) caching system. It extends the `BaseCaching` class and includes the `put` method to add items to the cache and the `get` method to retrieve items by key. The `LFUCache` class maintains a maximum number of items in the cache and discards the least frequently used item when the cache is full.