feat: Implement seat reservation API endpoints
This commit adds the necessary code to implement seat reservation API endpoints in the Express application. It includes routes for checking the number of available seats and reserving a seat. The Redis client is used to store and retrieve the current number of available seats. This feature will allow users to view the number of available seats and reserve a seat if it is available.